Handle LD and ST operations
Note that these aren't handled correctly: they're allowed to be placed
on any tile.  With any luck this will be OK because the placement tool
will constrain them to be on the same tile, but this should probably
be verified at some point.

Unfortunately this is somewhat difficult to do because you'd have to
walk the graph a bit and make sure that the memory whose address is
used to generate this offset is on the same tile as the memory
instruction.