Set tile owner from placement in files
This allows me to actually use a pflo file correctly.  The general
idea here is that I want to actually have the node constructor fully
construct a node.  This means that if a pflo file specifies the
placement of a node then we won't just set that string inside the node
but will actually look up the cooresponding node in the OCN graph and
store it.  This way there's only one place I need to check if a node
has been placed.

Note that this requires a change to libflo:

    8a35918 Allow users to specify a node creation function